[chore] #133 - store의 dto 패키지 request, response로 분리
