+++ src/main/java/com/napzak/domain/store/api/controller/StoreController.java
@@ -19,12 +19,12 @@ import com.napzak.domain.genre.api.dto.response.GenreNameDto;
 import com.napzak.domain.genre.api.dto.response.GenreNameListResponse;
 import com.napzak.domain.genre.api.exception.GenreErrorCode;
 import com.napzak.domain.store.api.StoreGenreFacade;
-import com.napzak.domain.store.api.dto.GenrePreferenceDto;
-import com.napzak.domain.store.api.dto.GenrePreferenceRequest;
-import com.napzak.domain.store.api.dto.LoginSuccessResponse;
-import com.napzak.domain.store.api.dto.MyPageResponse;
-import com.napzak.domain.store.api.dto.StoreInfoResponse;
-import com.napzak.domain.store.api.dto.StoreLoginResponse;
+import com.napzak.domain.store.api.dto.request.GenrePreferenceRequest;
+import com.napzak.domain.store.api.dto.response.GenrePreferenceDto;
+import com.napzak.domain.store.api.dto.response.LoginSuccessResponse;
+import com.napzak.domain.store.api.dto.response.MyPageResponse;
+import com.napzak.domain.store.api.dto.response.StoreInfoResponse;
+import com.napzak.domain.store.api.dto.response.StoreLoginResponse;
 import com.napzak.domain.store.api.exception.StoreErrorCode;
 import com.napzak.domain.store.api.exception.StoreSuccessCode;
 import com.napzak.domain.store.api.service.LoginService;
