--- src/main/java/com/napzak/domain/store/api/dto/StoreLoginResponse.java
@@ -1,17 +1,0 @@
-package com.napzak.domain.store.api.dto;
-
-import com.napzak.domain.store.core.entity.enums.Role;
-
-public record StoreLoginResponse(
-        String accessToken,
-        String nickname,
-        Role role
-) {
-    public static StoreLoginResponse of(
-            final String accessToken,
-            String nickname,
-            final Role role
-    ) {
-        return new StoreLoginResponse(accessToken, nickname, role);
-    }
-}
