+++ src/main/java/com/napzak/domain/store/api/dto/response/MyPageResponse.java
@@ -1,4 +1,4 @@
-package com.napzak.domain.store.api.dto;
+package com.napzak.domain.store.api.dto.response;
 
 public record MyPageResponse(
 	Long storeId,
